requirements.txt
Django>=3.0,<4.0
djangorestframework==3.13.1
mysqlclient==2.1.0
django-mysql==4.5.0